templates/invoice_extraction/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-extraction"
version = "0.1.0"
description = "A workflow that, given an invoice, extracts several key details using LlamaExtract"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "llama-index-workflows",
  "llama-cloud-services"
]

[tool.llamadeploy.workflows]
default = "invoice_extraction.workflow:workflow"
